Updated desc
Removed "Theme by" bit to keep consistent with all other themes. More changes will need to be made later to this code, as the option to evolve can happen when round%3 === 2, however the player has to choose whether to do this or attack, meaning the option to evolve should be a choice along with attacking when round%3 === 2 in the onNextRound() function. This would likely eliminate the need for the beforeNextRound() function, but I am not sure how to make that happen yet...